requirements.txt
requests==2.31.0
oauthlib==3.2.2
flask==2.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
ics==0.7.2